+++ src/main/java/io/xlorey/FluxLoader/client/core/Core.java
@@ -24,7 +24,7 @@ public class Core {
 
         EventManager.subscribe(new EventsHandler());
 
-        PluginManager.loadPlugins();
+        PluginManager.loadPlugins(true);
 
         WidgetManager.init();
     }